Add RecordingsMetadataManager service
